third_party/blink/web_tests/webmidi/requestmidiaccess-fail-expected.txt

CONSOLE WARNING: line 33: Web MIDI will ask a permission to use even if the sysex is not specified in the MIDIOptions since M77, around September 2019. See https://www.chromestatus.com/feature/5138066234671104 for more details.
Tests navigator.requestMIDIAccess failure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PermissionsHelper.setPermission is defined.
PASS errorName is 'InvalidStateError'
PASS requestMIDIAccess() fail as expected.
PASS successfullyParsed is true

TEST COMPLETE
